Moving all utils to sdv.cag._utils
