+++ sdv/cag/inequality.py
@@ -13,9 +13,6 @@ from sdv.cag._utils import (
     _validate_columns_not_primary_key,
     _validate_table_and_column_names,
     _validate_table_name_if_defined,
-)
-from sdv.cag.base import BaseConstraint
-from sdv.cag.utils import (
     _warn_if_timezone_aware_formats,
     cast_to_datetime64,
     compute_nans_column,
@@ -23,6 +20,7 @@ from sdv.cag.utils import (
     match_datetime_precision,
     revert_nans_columns,
 )
+from sdv.cag.base import BaseConstraint
 
 
 class Inequality(BaseConstraint):
